requirements.txt
# Core data & ML
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
optuna>=3.5.0
shap>=0.44.0

# Regime & volatility
hmmlearn>=0.3.0
arch>=6.3.0

# FRED API
fredapi>=0.5.1

# Utilities
pyyaml>=6.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# API/Deployment
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
requests>=2.31.0

# Testing
pytest>=7.4.0
